test(streambot): split the test directory into sub-domains
Why:
`packages/streambot/test` held 57 test files flat. It is the only directory in
this series that is over on the test budget rather than the source one.

What:
57 of 59 direct children move into `media/` (22), `voice/` (16), `session/`
(12) and `unit/` (6).

Two files stay at the root, and both would fail quietly if they moved:
`setup.ts` is the suite's setupFile, named by path in the shared vitest config;
`architecture-boundaries.test.ts` derives its package root with
`import.meta.dir.replace(/\/test$/)`, so from a sub-domain it would have
cruised the wrong root and reported a clean architecture for a directory it
never inspected.

`support/fake-realtime-transport.ts` moves in beside its only consumer.
Reaching it from `voice/` would have needed a relative parent import, which
`custom-rules/no-parent-imports` forbids, and streambot's `imports` map has no
wildcard to express it as a package import.

The 400 `.dopus` voice-corpus fixtures are deliberately untouched: they are not
code files, they do not count toward either budget, and
`packages/streambot/Dockerfile` excludes them from the image by exact path. A
stale exclude there fails open - it would silently ship a multi-megabyte corpus
- so leaving the fixtures in place is the safe outcome, not an oversight.

Verification:
- `turbo run test --filter=@shepherdjerred/streambot` - 57 files, 608 tests passed
- `turbo run typecheck lint --filter=@shepherdjerred/streambot` - clean;
  architecture 110 modules clean, which is the check that would have gone
  vacuous had the boundaries test moved
- every `import.meta`-relative path in the directory re-resolved against disk
- the Dockerfile's fixture exclude still matches an existing path
- `jscpd-check` ratchet passed, 436 pairs preserved; suppressions clean
