+++ packages/streambot/test/media/library.test.ts
