+++ packages/streambot/test/session/command-handler.test.ts
