+++ packages/streambot/test/session/permissions.test.ts
