+++ packages/streambot/test/session/session-manager.test.ts
