+++ packages/streambot/test/unit/config.test.ts
