+++ packages/streambot/test/voice/voice-model-assets.test.ts
